% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBA_deg.R
\name{BBA_deg}
\alias{BBA_deg}
\title{deg data for example black browed albatross foraging trip.}
\format{A data frame with 782 rows and 3 variables:
\describe{
  \item{dtime}{datetime object as POSIXct in UTC}
  \item{duration}{duration of wet or dry state in sec}
  \item{wet.dry}{wet or dry state}
}}
\description{
deg data for example black browed albatross foraging trip during the summer solstice.
}

