% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBA_lux.R
\name{BBA_lux}
\alias{BBA_lux}
\title{lux data for example black browed albatross foraging trip.}
\format{
A data frame with 2484 rows and 2 variables:
\describe{
  \item{dtime}{datetime object as POSIXct in UTC}
  \item{lig}{light intensity}
}
}
\description{
lux data for example black browed albatross foraging trip during the summer solstice.
}
