% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBA_sst.R
\name{BBA_sst}
\alias{BBA_sst}
\title{sst data for one black browed albatross foraging trip.}
\format{A data frame with 25 rows and 3 variables:
\describe{
  \item{dtime}{datetime object as POSIXct in UTC}
  \item{date}{date object as Date object}
  \item{temp}{median recorded water temperature in degree Celsius}
}}
\description{
sst data for one black browed albatross foraging trip.
}

