% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mcp.R
\name{calc_mcp}
\alias{calc_mcp}
\title{Calculate MCPs}
\usage{
calc_mcp(xy, percent = 100)
}
\arguments{
\item{xy}{SF POINTS data.frame with one column denoting polygon ID.}

\item{percent}{Percentage considered to determine MCP}

\item{prob1}{lower quantile considered}

\item{prob2}{upper quantile considered}
}
\description{
This function calculates minimum convex polygons (MCP). It is based on the mcp function from adehabitatHR.
It does not calculate area and does not require a minimum number of points. It is independent of adehabitatHR and sp, but relies on sf.
}
