% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_data.frame.R
\name{f}
\alias{f}
\title{transform list into data frame}
\usage{
f(x)
}
\arguments{
\item{x}{list to be tranformed}
}
\value{
A data.frame for further use in ProMM
}
\description{
transform list into data frame
}

