% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fun_list_to_dataframe}
\alias{fun_list_to_dataframe}
\title{Internal function to collapse lists into data.frame}
\usage{
fun_list_to_dataframe(x)
}
\arguments{
\item{x}{List that should be collapsed into a data.frame}
}
\description{
Internal function to collapse lists into data.frame
}
\keyword{internal}
