% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fun_mcp}
\alias{fun_mcp}
\title{Internal function to calculate minimum convex polygon (MCP)}
\usage{
fun_mcp(xy, percent = 100)
}
\arguments{
\item{xy}{data in x dimension to be plotted}

\item{percent}{Percentage considered to determine MCP}

\item{prob1}{lower quantile considered}

\item{prob2}{upper quantile considered}
}
\description{
This internal function calculates minimum convex polygons (MCP) without calculating area and without minimum required points to be independent of adehabitatHR.
}
