% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fun_time_dry}
\alias{fun_time_dry}
\title{Internal function to calculate the proportion of time dry}
\usage{
fun_time_dry(x, y, time_max, wr)
}
\arguments{
\item{x}{point cloud data.frame}

\item{y}{act data.frame given to prob_algorithm}

\item{time_max}{End time point}

\item{wr}{sampling rate of conductivity switch in sec (e.g. MK15 & MK3006 sample every 3 sec) as given to prob_algorithm}
}
\description{
Internal function to calculate the proportion of time between to twilight events a logger was registered as being dry
}
\keyword{internal}
