% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fun_weight_ice}
\alias{fun_weight_ice}
\title{Internal function to calculate the weight of each particle given sea ice concentration}
\usage{
fun_weight_ice(x, y)
}
\arguments{
\item{x}{point cloud data.frame}

\item{y}{ice.conc.cutoff given to prob_algorithm}
}
\description{
Internal function to calculate the weight of each particle given sea ice concentration
}
\keyword{internal}
