% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fun_weight_speed}
\alias{fun_weight_speed}
\title{Internal function to calculate the weight of each particle given the animals speed}
\usage{
fun_weight_speed(x, spd, spw)
}
\arguments{
\item{x}{point cloud data.frame}

\item{spd}{speed.dry given to prob_algorithm}

\item{spw}{speed.wet given to prob_algorithm}
}
\description{
Internal function to calculate the weight of each particle given the animals speed
}
