% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fun_weight_sst}
\alias{fun_weight_sst}
\title{Internal function to calculate the weight of each particle given SST and tag temperature}
\usage{
fun_weight_sst(x, y, z)
}
\arguments{
\item{x}{point cloud data.frame}

\item{y}{sst.sd given to prob_algorithm}

\item{z}{max.sst.diff given to prob_algorithm}
}
\description{
Internal function to calculate the weight of each particle given SST and tag temperature
}
