% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipe_to_dataframe.R
\name{ipe_to_dataframe}
\alias{ipe_to_dataframe}
\title{transform .ipe file into data frame}
\usage{
ipe_to_dataframe(ipefile)
}
\arguments{
\item{ipefile}{the ipe file to be transformed}
}
\value{
A data.frame for further use in GeoLight or ProMM

modified after trnTrans() in GeoLight by Tamara Emmenegger
}
\description{
transform migrate tech .ipe files into a data frame with tFirst, tSecond and type
}

