% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.NOAA.OISST.V2.R
\name{load.NOAA.OISST.V2}
\alias{load.NOAA.OISST.V2}
\title{load NOAA OISST V2}
\usage{
load.NOAA.OISST.V2(fname, lsmask, lonW, lonE, latS, latN, date1, date2,
  use.landmask = F, extract.value = "sst")
}
\arguments{
\item{fname}{full path to NetCDF data file}

\item{lsmask}{full path to land-sea mask NetCDF file}

\item{lonW}{western-most longitude of search area, must be smaller than lonE}

\item{lonE}{eastern-most longitude of search area, must be larger than lon1}

\item{latS}{southern-most latitude of search area, must be smaller than latN}

\item{latN}{northern-most latitude of search area, must be larger than latS}

\item{date1}{first date in file to extract, must be Date class}

\item{date2}{last date in file to extract, must be Date class}

\item{use.landmask}{use land mask TRUE or FALSE}

\item{extract.value}{which data to extract: "sst" - SST, "err" - SST error, "icec" - sea ice concentration}
}
\value{
A 3-dimensional array with latitudes in rows, longitudes in columns, and dates along the 3rd dimension. The value [1,1,1] is the northernmost, westernmost lat/long location on the 1st date. The value [1,1,2] is the 2nd date at the same lat/long location (if more than 1 date is requested).

To extract lat/lon/date values from the output array, use the dimnames() function:

lats = as.numeric(dimnames(sst2)$Lat)

lons = as.numeric(dimnames(sst2)$Long)

dates = as.Date(dimnames(sst2)$Date)



NetCDF files should be downloaded from the links on:

http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.highres.html

In addition to the temperature data files, also download a copy of the landmask file lsmask.oisst.v2.nc from the same page. Inside the NetCDF files, data are available on a 0.25 degree latitude x 0.25 degree longitude global grid (720x1440 cells) From -89.875N to 89.875N, 0.125E to 359.875E. Locations are at the CENTER of a grid cell.

modified after Luke Miller Nov 25, 2014; https://github.com/millerlp/Misc_R_scripts/blob/master/NOAA_OISST_ncdf4.R
}
\description{
This function takes 1-year-long NetCDF files from the
ftp://ftp.cdc.noaa.gov/Datasets/noaa.oisst.v2.highres/ directory
}

