% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_NOAA_OISST_V2.R
\name{load_NOAA_OISST_V2}
\alias{load_NOAA_OISST_V2}
\title{load NOAA OISST V2}
\usage{
load_NOAA_OISST_V2(FILE_NAME, LONS, LATS, DATE, extract.value = "sst")
}
\arguments{
\item{FILE_NAME}{full path to NetCDF data file}

\item{LONS}{Point longitudes}

\item{LATS}{Point latitudes}

\item{DATE}{date considered}

\item{extract.value}{which data to extract: "sst" - SST, "err" - SST error, "icec" - sea ice concentration}
}
\value{
A vector of values for each lon lat pair given the extract.value

NetCDF files should be downloaded from the links on:

http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.highres.html

In addition to the temperature data files, also download a copy of the landmask file lsmask.oisst.v2.nc from the same page. Inside the NetCDF files, data are available on a 0.25 degree latitude x 0.25 degree longitude global grid (720x1440 cells) From -89.875N to 89.875N, 0.125E to 359.875E. Locations are at the CENTER of a grid cell.
}
\description{
This function extracts information from NOAA OISST V2 NetCDF files available at
ftp://ftp.cdc.noaa.gov/Datasets/noaa.oisst.v2.highres/ directory
}
