% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotek_to_dataframe.R
\name{lotek_to_dataframe}
\alias{lotek_to_dataframe}
\title{transform lotek day log data into data frame}
\usage{
lotek_to_dataframe(date, sunrise, sunset, TRLon, TRLat)
}
\arguments{
\item{date}{column "TimeS" in daylog file as as.Date()}

\item{sunrise}{column "Sunrise" in daylog file as as.character()}

\item{sunset}{column "Sunset" in daylog file as as.character()}

\item{TRLon}{threshold method calculated longitude in daylog file}

\item{TRLat}{threshold method calculated latitude in daylog file}
}
\value{
A data.frame for further use in GeoLight or probGLS

modified after trnTrans() in GeoLight by Tamara Emmenegger
}
\description{
transform lotek daylog file data into a data frame with tFirst, tSecond and type
}

