% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{plot map of algorithm output}
\usage{
plot_map(pr, legend.position = "topleft")
}
\arguments{
\item{pr}{probGLS algorithm output}

\item{legend.position}{position of colour legend. Options are "topleft", "topright", "bottomleft", and "bottomright"}
}
\description{
plot map of algorithm output
}
