% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{poly_frame}
\alias{poly_frame}
\title{Internal function to plot polygon shapes in plot_timeline}
\usage{
poly_frame(data1, data2, prob1, prob2)
}
\arguments{
\item{data1}{data in x dimension to be plotted}

\item{data2}{data in y dimension to be plotted}

\item{prob1}{lower quantile considered}

\item{prob2}{upper quantile considered}
}
\description{
This internal function determines polygon shapes to plot in plot_timeline
}
