% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_algorithm.R
\name{prob_algorithm}
\alias{prob_algorithm}
\title{probabilistic algorithm for geolocation data}
\usage{
prob_algorithm(particle.number = 2000, iteration.number = 60,
  loess.quartile = NULL, tagging.location = c(0, 0), tagging.date,
  retrieval.date, sunrise.sd = c(2.49, 0.94, 4.98), sunset.sd = c(2.49,
  0.94, 4.98), range.solar = c(-7, -1), speed.wet = c(20, 0.2, 25),
  speed.dry = c(20, 0.2, 25), sst.sd = 0.5, max.sst.diff = 3,
  days.around.spring.equinox = c(10, 10), days.around.fall.equinox = c(10,
  10), ice.conc.cutoff = 1, boundary.box = c(-180, 180, -90, 90),
  med.sea = T, black.sea = T, baltic.sea = T, caspian.sea = T,
  land.mask = NULL, east.west.comp = T, sensor, trn, act,
  wetdry.resolution = 30, backward = F,
  NOAA.OI.location = "E:/environmental data/SST/NOAA OI SST V2")
}
\arguments{
\item{particle.number}{number of particles for each location cloud used in the model}

\item{iteration.number}{number of iterations}

\item{loess.quartile}{quartiles for loessFilter (GeoLight), if NULL loess filter is not used}

\item{tagging.location}{tagging location longitude and latitude}

\item{tagging.date}{deployment data as POSIXct or Date object}

\item{retrieval.date}{retrieval date as POSIXct or Date object}

\item{sunrise.sd}{output vector from twilight_error_estimation}

\item{sunset.sd}{output vector from twilight_error_estimation}

\item{range.solar}{min and max of solar angle range in degree}

\item{speed.wet}{optimal speed, speed standard deviation and max speed allowed if logger is wet in m/s}

\item{speed.dry}{optimal speed, speed standard deviation and max speed allowed if logger is dry in m/s}

\item{sst.sd}{SST standard deviation in degree C}

\item{max.sst.diff}{max difference in SST allowed in degree C}

\item{days.around.spring.equinox}{days before the Spring equinox and days after the Spring equinox. The Spring equinox is assumed constant at 20 March.}

\item{days.around.fall.equinox}{days before the Fall equinox, days after the Fall equinox. The Fall equinox is assumed constant at 22 September.}

\item{ice.conc.cutoff}{max percentage of sea ice in which the animal is believed to be}

\item{boundary.box}{min lon, max lon, min lat and max lat of extrem boundary where you expect an animal to be}

\item{med.sea}{if T classifiy mediterranean sea as land}

\item{black.sea}{if T classifiy black sea as land}

\item{baltic.sea}{if T classifiy baltic sea as land}

\item{caspian.sea}{if T classifiy caspian sea as land}

\item{land.mask}{if T animal is only using ocean areas, if F animal is only using land areas, if NULL no land mask used}

\item{east.west.comp}{if T apply biotrack east west movement compensation (Biotrack manual v11 page 31pp.)}

\item{sensor}{data.frame with daily SST data deduced from tag temperature readings (sst_deduction ouput), NULLif no SST data is available (SST will not be used)}

\item{trn}{data.frame containing twilights and at least tFirst, tSecond and type (same as computed by trn_to_dataframe, ipe_to_dataframe or lotek_to_dataframe)}

\item{act}{data.frame containing wet dry data (e.g. .act file from Biotrack loggers or .deg file from migrate tech loggers), NULL if no wetdry data is available (algorithm will assume that the logger was always dry)}

\item{wetdry.resolution}{sampling rate of conductivity switch in sec (e.g. MK15 & MK3006 sample every 3 sec)}

\item{backward}{run algorithm from end to start}

\item{NOAA.OI.location}{directory location of NOAA OI V2 NCDF files as well as land mask file 'lsmask.oisst.v2.nc' (downloadable from http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.highres.html)}
}
\value{
A list with: [1] all positions, [2] geographic median positions, [3] all possible particles, [4] input parameters, [5] model run time. List items 1 to 3 are returned as SpatialPointsDataframe.
}
\description{
prob_algorithm is a simple probabilistic algorithm to be used with geolocation data.
}
\details{
Many weighting parameters can be used. Some others (which are not yet implemented) are: surface air temperature and topography/ bathymetry.
}
\examples{
######################################
# example black browed albatross track 
######################################

# define start and end datetimes ----
start <- as.POSIXct("2014-12-13 17:55", tz="UTC")
end   <- as.POSIXct("2014-12-22 08:55", tz="UTC")

# light data ----
trn           <- twilightCalc(BBA_lux$dtime, BBA_lux$lig, ask = FALSE, LightThreshold = 2)

# sst data ----
sen           <- sst_deduction(datetime = BBA_sst$dtime, temp = BBA_sst$temp, temp.range = c(-2,30))

# wet dry data ----
act           <- BBA_deg[BBA_deg$wet.dry=="wet",]
act$wetdry    <- act$duration

# twilight error distribution estimation ----
tw            <- twilight_error_estimation()

# download environmental data ----

# download yearly NetCDF files for (replace YEAR with appropriate number): 
# daily mean SST                   -> 'sst.day.mean.YEAR.v2.nc'
# daily SST error                  -> 'sst.day.err.YEAR.v2.nc'
# daily mean sea ice concentration -> 'icec.day.mean.YEAR.v2.nc'

# from:
# http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.highres.html
# and place all into the same folder

# Also, download the land mask file: 'lsmask.oisst.v2.nc' from the same directory 
# and place it in the same folder as all the other NetCDF files

# run algorithm ----
pr   <- prob_algorithm(trn                         = trn, 
                      sensor                      = sen, 
                      act                         = act, 
                      tagging.date                = start, 
                      retrieval.date              = end, 
                      loess.quartile              = NULL, 
                      tagging.location            = c(-36.816,-54.316), 
                      particle.number             = 1000, 
                      iteration.number            = 100,
                      sunrise.sd                  = tw,
                      sunset.sd                   = tw,
                      range.solar                 = c(-7,-1),
                      boundary.box                = c(-120,40,-90,0),
                      days.around.spring.equinox  = c(0,0), 
                      days.around.fall.equinox    = c(0,0),
                      speed.dry                   = c(12,6,45),
                      speed.wet                   = c(1,1.3,5), 
                      sst.sd                      = 0.5,       
                      max.sst.diff                = 3,          
                      east.west.comp              = T,
                      land.mask                   = T, 
                      ice.conc.cutoff             = 1, 
                      wetdry.resolution           = 1,
                      NOAA.OI.location            = "folder with environmental data and land mask")

# plot lat, lon, SST vs time ----
plot_timeline(pr,degElevation = NULL)

# plot lon vs lat map ----
plot_map(pr)
}

