% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_algorithm_exp.R
\name{prob_algorithm_exp}
\alias{prob_algorithm_exp}
\title{probabilistic algorithm for geolocation data (experimental version)
- currently trying to implement conductivity data for baltic eiders}
\usage{
prob_algorithm_exp(particle.number = 2000, iteration.number = 60,
  loess.quartile = NULL, tagging.location = c(0, 0), tagging.date,
  retrieval.date, sunrise.sd = c(2.49, 0.94, 4.98), sunset.sd = c(2.49,
  0.94, 4.98), range.solar = c(-7, -1), speed.wet = c(20, 0.2, 25),
  speed.dry = c(20, 0.2, 25), sst.sd = 0.5, max.sst.diff = 3,
  days.around.spring.equinox = c(10, 10),
  days.around.fall.equinox = c(10, 10), ice.conc.cutoff = 1,
  boundary.box = c(-180, 180, -90, 90), med.sea = T, black.sea = T,
  caspian.sea = T, land.mask = T, east.west.comp = T, sensor, trn,
  act, cond, cond.threshold = c(115, 90), wetdry.resolution = 30,
  backward = F,
  NOAA.OI.location = "E:/environmental data/SST/NOAA OI SST V2")
}
\arguments{
\item{particle.number}{number of particles for each location cloud used in the model}

\item{iteration.number}{number of iterations}

\item{loess.quartile}{quartiles for loessFilter (GeoLight), if NULL loess filter is not used}

\item{tagging.location}{tagging location longitude and latitude}

\item{tagging.date}{deployment data as POSIXct or Date object}

\item{retrieval.date}{retrieval date as POSIXct or Date object}

\item{sunrise.sd}{output vector from twilight_error_estimation}

\item{sunset.sd}{output vector from twilight_error_estimation}

\item{range.solar}{min and max of solar angle range in degree}

\item{speed.wet}{optimal speed, speed standard deviation and max speed allowed if logger is wet in m/s}

\item{speed.dry}{optimal speed, speed standard deviation and max speed allowed if logger is dry in m/s}

\item{sst.sd}{SST standard deviation in degree C}

\item{max.sst.diff}{max difference in SST allowed in degree C}

\item{days.around.spring.equinox}{days before the Spring equinox and days after the Spring equinox. The Spring equinox is assumed constant at 20 March.}

\item{days.around.fall.equinox}{days before the Fall equinox, days after the Fall equinox. The Fall equinox is assumed constant at 22 September.}

\item{ice.conc.cutoff}{max percentage of sea ice in which the animal is believed to be}

\item{boundary.box}{min lon, max lon, min lat and max lat of extrem boundary where you expect an animal to be}

\item{med.sea}{if T classifiy mediterranean sea as land}

\item{black.sea}{if T classifiy black sea as land}

\item{caspian.sea}{if T classifiy caspian sea as land}

\item{land.mask}{if T animal is only using ocean areas, if F animal is only using land areas, if NULL no land mask used}

\item{east.west.comp}{if T apply biotrack east west movement compensation (Biotrack manual v11 page 31pp.)}

\item{sensor}{data.frame with daily SST data deduced from tag temperature readings (sst_deduction ouput), NULLif no SST data is available (SST will not be used)}

\item{trn}{data.frame containing twilights and at least tFirst, tSecond and type (same as computed by trn_to_dataframe, ipe_to_dataframe or lotek_to_dataframe)}

\item{act}{data.frame containing wet dry data (e.g. .act file from Biotrack loggers or .deg file from migrate tech loggers), NULL if no wetdry data is available (algorithm will assume that the logger was always dry)}

\item{cond}{data.frame containing raw conductivity data , NULL if no conductivity data is available (algorithm will assume that the logger was always in water and not in the Baltic Sea)}

\item{cond.threshold}{expected two values: first values = upper threshold, i.e. not in the Baltic Sea; second value = lower threshold, i.e. animal on land below this threshold; animal assume to be in the Baltic Sea (i.e. within the box: lon>14 & lon<33.5 & lat>51.4 & lat<66.2) inbetween the two thresholds}

\item{wetdry.resolution}{sampling rate of conductivity switch in sec (e.g. MK15 & MK3006 sample every 3 sec)}

\item{backward}{run algorithm from end to start}

\item{NOAA.OI.location}{directory location of NOAA OI V2 NCDF files as well as land mask file 'lsmask.oisst.v2.nc' (downloadable from http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.highres.html)}

\item{baltic.sea}{if T classifiy baltic sea as land}
}
\value{
A list with: [1] all positions, [2] geographic median positions, [3] all possible particles, [4] input parameters, [5] model run time. List items 1 to 3 are returned as SpatialPointsDataframe.
}
\description{
prob_algorithm is a simple probabilistic algorithm to be used with geolocation data.
}
\details{
Many weighting parameters can be used. Some others (which are not yet implemented) are: surface air temperature and topography/ bathymetry.
}
