% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trn_to_dataframe.r
\name{trn_to_dataframe}
\alias{trn_to_dataframe}
\title{transform .trn file into data frame}
\usage{
trn_to_dataframe(trnfile)
}
\arguments{
\item{trnfile}{the trn file to be transformed}
}
\value{
A data.frame for further use in GeoLight or ProMM

modified after trnTrans() in GeoLight by Tamara Emmenegger
}
\description{
transform biotrack or BAS .trn files into a data frame with tFirst, tSecond and type
}

