% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilight_error_estimation.R
\name{twilight_error_estimation}
\alias{twilight_error_estimation}
\title{Estimate the assumed error around a twilight event}
\usage{
twilight_error_estimation(shape = 2.49, scale = 0.94, delay = 0)
}
\arguments{
\item{shape}{shape parameter of twilight error (in minutes)}

\item{scale}{scale parameter of twilight error (in minutes)}

\item{delay}{estimated delay of observed twilight event compared to true twilight event (in minutes)}
}
\description{
This function is a visual aid to estimate the assumed error around a twilight event. The determined shape and scale parameters will be used as input for the prob_algorithm function. This function assumes that the error around a twilight event is log normal distributed.
}
\details{
This function can be used to visualise the error distribution around each twilight event which will be used in the prob_algorithm function. The error is assumed to be log normal distributed. In the case of sunrises, only a small error before and a higher possible error with a long tail after the twilight event is assumed (this is reversed in the case of sunsets).
}

