% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{plot map of algorithm output}
\usage{
plot_map(pr, legend.position = "topleft")
}
\arguments{
\item{pr}{probGLS algorithm output}

\item{legend.position}{position of colour legend. Options are "topleft", "topright", "bottomleft", and "bottomright"}
}
\description{
plot map of algorithm output
}
\details{
This function plots the most probable path constructed using prob_algorithm and associated uncertainty.

The locations are colour coded by season. First and last location are connected by stippled lines to the capture location (assuming that it is also the recapture location). Dotted shapes denote 95% minimum convex polygon (MCP) of all estimated particles in the track. Increasing grey-scaled shapes illustrate 95, 75, 50 and 25% MCPs of all particles.
}
