% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{plot timeline of model output}
\usage{
plot_timeline(pr, degElevation = NULL)
}
\arguments{
\item{pr}{promm model output}

\item{degElevation}{sun elevation angle for comparison threshold method, if NULL no threshold positions are plotted}
}
\description{
plot timeline of model output
}

