% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{plot timeline of algorithm output}
\usage{
plot_timeline(pr, solar.angle = NULL, center.longitude = 0)
}
\arguments{
\item{pr}{probGLS algorithm output}

\item{solar.angle}{solar angle to compare prob_algorithm output with the classical threshold method, if NULL no threshold positions are plotted}

\item{center.longitude}{around which longitude should the plot be centered? Only 0 and 180 implemented}
}
\description{
plot timeline of algorithm output
}
