% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promm.R
\name{promm}
\alias{promm}
\title{probabalistic movement model for geolocation data}
\usage{
promm(particle.number = 500, bootstrap.number = 50, loess.quartile = NULL,
  colony = c(0, 0), tagging.date, retrieval.date, twilight.sd = 10,
  range.sun.elev = c(-7, -1, 0.1), speed.wet = c(20, 0.2, 25),
  speed.dry = c(20, 0.2, 25), sst.sd = 0.5, max.sst.diff = 3,
  days.around.spring.equinox = c(10, 10), days.around.fall.equinox = c(10,
  10), ice.conc.cutoff = 0.9, boundary.box = c(-90, 70, 0, 90),
  med.black.sea = T, baltic.sea = T, caspian.sea = T,
  east.west.comp = T, sensor, trn, act,
  NOAA.OI.location = "E:/environmental data/SST/NOAA OI SST V2")
}
\arguments{
\item{particle.number}{number of particles for each location cloud used in the model}

\item{bootstrap.number}{number of iterations}

\item{loess.quartile}{quartiles for loess.filter (GeoLight), if NULL then loess filter is not used}

\item{colony}{colony longitude and latitude}

\item{tagging.date}{deployment data}

\item{retrieval.date}{retrieval date}

\item{twilight.sd}{sd around each unsrise or sunset event in min}

\item{range.sun.elev}{min and max sun angle in degree, resolution of sun elevation angle}

\item{speed.wet}{optimal speed, speed sd and max speed allowed if logger is wet in m/s}

\item{speed.dry}{optimal speed, speed sd and max speed allowed if logger is dry in m/s}

\item{sst.sd}{SST sd in degree C}

\item{max.sst.diff}{max difference in SST allowed in degree C}

\item{days.around.spring.equinox}{days before the Spring equinox and days after the Spring equinox, the Spring equinox is assumed constant at 20 March}

\item{days.around.fall.equinox}{days before the Fall equinox, days after the Fall equinox, the Fall equinox is assumed constant at 22 September}

\item{ice.conc.cutoff}{min percentage of sea ice in which there are no birds}

\item{boundary.box}{min lon, max lon, min lat and max lat of extrem boundary where you expect an animal to be}

\item{med.black.sea}{if T remove meditearanian and black sea}

\item{baltic.sea}{if T remove baltic sea}

\item{caspian.sea}{if T remove caspian sea}

\item{east.west.comp}{if T apply biotrack east west movement compensation}

\item{sensor}{sensor data input}

\item{trn}{sunrise and sunset data input with the same structure as computed by trn_to_dataframe or ipe_to_dataframe}

\item{act}{wet dry data input}

\item{NOAA.OI.location}{directory location of NOAA OI V2 files as well as land mask file 'lsmask.oisst.v2.nc'}

\item{plot.it}{plot each step}
}
\value{
A list of modeling results: [1] all bootstrapped positions, [2] geographic median positions, [3] all possible particles, [4] time taken
}
\description{
modeling stuff
}

