% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_deduction.R
\name{sst_deduction}
\alias{sst_deduction}
\title{SST deduction}
\usage{
sst_deduction(datetime, temp, temp.range = c(-2, 30))
}
\arguments{
\item{datetime}{date time object as POSIXct in UTC}

\item{temp}{temperature readings}

\item{temp.range}{min and max of temperature range}
}
\description{
SST deduction from logger temperature reading
}

